Corrección en AutoMapper

+++ ApiLoginJdepaz.Core/Domains/Productos/ProductoResponse.cs
@@ -9,7 +9,7 @@ namespace ApiLoginJdepaz.Core.Domains.Productos
         public string SKU { get; set; }
         public string nombre { get; set; }
         public int cantidad { get; set; }
-        public decimal precio { get; set; }
+        public double precio { get; set; }
         public string descripcion { get; set; }
         public string imagen { get; set; }
     }
